requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
docstring_parser==0.17.0
fastapi==0.116.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.111.0
google-cloud-bigquery==3.36.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai>==1.32.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic>==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
